requirements.txt
kivy==2.3.1
pyjnius
plyer
SpeechRecognition
pyttsx3
Cython==0.29.33 